remove comment and rename file #2919
