+++ assets/app-back-bo.ts
@@ -11,12 +11,12 @@ import './controllers/form_account';
 import './controllers/form_nde';
 import './controllers/form_notification';
 import './controllers/form_visite';
+import './controllers/form_modal_handler';
 
 import './controllers/back_archived_signalements/back_archived_signalements_reactiver'
 import './controllers/back_auto_affectation_rule/form_auto_affectation_rule'
 import './controllers/back_partner_view/form_partner';
 import './controllers/back_signalement_view/back_view_signalement';
-import './controllers/back_signalement_view/form_edit_modal';
 import './controllers/back_signalement_view/form_upload_documents';
 import './controllers/back_signalement_view/input_autocomplete_bailleur';
 import './controllers/back_signalement_view/form_cloture_modal';
